candidate_mining/test/requirements.txt
absl-py==0.4.1
appnope==0.1.0
asn1crypto==0.24.0
astor==0.7.1
backcall==0.1.0
bert-pytorch==0.0.1a2
bleach==2.1.4
boto==2.49.0
boto3==1.9.46
botocore==1.12.46
bottle==0.12.13
bz2file==0.98
certifi==2018.10.15
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
cycler==0.10.0
cymem==1.31.2
cytoolz==0.9.0.1
decorator==4.3.0
dill==0.2.8.2
docutils==0.14
entrypoints==0.2.3
fire==0.1.3
gast==0.2.0
gensim==3.6.0
grpcio==1.16.1
grpcio-tools==1.12.1
h5py==2.8.0
html5lib==1.0.1
idna==2.7
ipykernel==4.9.0
ipython==6.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.1
jedi==0.12.1
jieba==0.39
Jinja2==2.10
jmespath==0.9.3
JPype1==0.6.3
jsonschema==2.6.0
jupyter==1.0.0
jupyter-client==5.2.3
jupyter-console==5.2.0
jupyter-core==4.4.0
kaggle==1.5.0
Keras-Applications==1.0.6
Keras-Preprocessing==1.0.5
kiwisolver==1.0.1
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.3
mistune==0.8.3
mkl-fft==1.0.4
mkl-random==1.0.1
msgpack==0.5.6
msgpack-numpy==0.4.3.1
murmurhash==0.28.0
nbconvert==5.3.1
nbformat==4.4.0
nltk==3.3
nose-parameterized==0.6.0
notebook==5.6.0
numpy==1.15.1
pandas==0.23.4
pandocfilters==1.4.2
parameterized==0.6.1
parso==0.3.1
patsy==0.5.0
pexpect==4.6.0
pickleshare==0.7.4
plac==0.9.6
preshed==1.0.1
prometheus-client==0.3.1
prompt-toolkit==1.0.15
protobuf==3.6.1
ptyprocess==0.6.0
pycparser==2.18
Pygments==2.2.0
pygtrie==2.3
pyhanlp==0.1.44
pyOpenSSL==18.0.0
pyparsing==2.2.0
PySocks==1.6.8
python-dateutil==2.7.3
python-slugify==1.2.6
pytz==2018.5
pyzmq==17.1.2
qtconsole==4.4.1
regex==2018.7.11
requests==2.19.1
rouge==0.3.1
s3transfer==0.1.13
scipy==1.1.0
seaborn==0.9.0
Send2Trash==1.5.0
simplegeneric==0.8.1
six==1.11.0
smart-open==1.7.1
spacy==2.0.12
statsmodels==0.9.0
tensorboard==1.12.0
tensorflow==1.12.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
thinc==6.10.3
toolz==0.9.0
torch==0.4.1
tornado==5.1
tqdm==4.25.0
traitlets==4.3.2
ujson==1.35
Unidecode==1.0.22
urllib3==1.22
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.14.1
widgetsnbextension==3.4.1
wrapt==1.10.11
xlrd==1.1.0
